+++ src/Data/IntMap/Monoidal/Strict.hs
@@ -147,10 +147,15 @@ import qualified Data.IntMap.Strict as M
 import Control.Lens
 import Control.Newtype
 import Data.Aeson(FromJSON, ToJSON, FromJSON1, ToJSON1)
-#if MIN_VERSION_containers(0,5,9)
 import Data.Functor.Classes
-#endif
 import Data.Align
+#ifdef MIN_VERSION_semialign
+import Data.Semialign (Unalign)
+#if MIN_VERSION_semialign(1,1,0)
+import Data.Zip (Zip)
+#endif
+#endif
+import qualified Witherable
 
 -- | An 'IntMap' with monoidal accumulation
 newtype MonoidalIntMap a = MonoidalIntMap { getMonoidalIntMap :: M.IntMap a }
@@ -159,15 +164,20 @@ newtype MonoidalIntMap a = MonoidalIntMap { getMonoidalIntMap :: M.IntMap a }
               FromJSON, ToJSON, FromJSON1, ToJSON1,
               Data, Typeable, Align
 #if MIN_VERSION_these(0,8,0)
-              , Semialign
+             , Semialign
 #endif
+#ifdef MIN_VERSION_semialign
+             , Unalign
+#if MIN_VERSION_semialign(1,1,0)
+             , Zip
+#endif
+#endif
+             , Witherable.Filterable
              )
 
-#if MIN_VERSION_containers(0,5,9)
 deriving instance Eq1 MonoidalIntMap
 deriving instance Ord1 MonoidalIntMap
 deriving instance Show1 MonoidalIntMap
-#endif
 
 type instance Index (MonoidalIntMap a) = Int
 type instance IxValue (MonoidalIntMap a) = a
@@ -234,6 +244,8 @@ instance Semigroup a => IsList.IsList (MonoidalIntMap a) where
     toList = M.toList . unpack
     {-# INLINE toList #-}
 #endif
+
+instance Witherable.Witherable MonoidalIntMap
 
 -- | /O(1)/. A map with a single element.
 singleton :: Int -> a -> MonoidalIntMap a